Refactor muicss react types per pull request feedback
