+++ types/muicss/lib/react/appbar.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Appbar extends React.Component<P.DivProps> {}
