+++ types/muicss/lib/react/button.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Button extends React.Component<P.ButtonProps> {}
