+++ types/muicss/lib/react/checkbox.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Checkbox extends React.Component<P.CheckboxProps> {}
