+++ types/muicss/lib/react/col.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Col extends React.Component<P.ColProps> {}
