+++ types/muicss/lib/react/container.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Container extends React.Component<P.ContainerProps> {}
