+++ types/muicss/lib/react/divider.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Divider extends React.Component<P.DividerProps> {}
