+++ types/muicss/lib/react/dropdown-item.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class DropdownItem extends React.Component<P.DropdownItemProps> {}
