+++ types/muicss/lib/react/dropdown.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Dropdown extends React.Component<P.DropdownProps> {}
