+++ types/muicss/lib/react/form.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Form extends React.Component<P.FormProps> {}
