+++ types/muicss/lib/react/input.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Input extends React.Component<P.InputProps> {}
