+++ types/muicss/lib/react/option.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Option extends React.Component<P.OptionProps> {}
