+++ types/muicss/lib/react/panel.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Panel extends React.Component<P.PanelProps> {}
