+++ types/muicss/lib/react/radio.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Radio extends React.Component<P.RadioProps> {}
