+++ types/muicss/lib/react/row.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Row extends React.Component<P.RowProps> {}
