+++ types/muicss/lib/react/select.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Select extends React.Component<P.SelectProps> {}
