+++ types/muicss/lib/react/tab.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Tab extends React.Component<P.TabProps> {}
