+++ types/muicss/lib/react/tabs.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class Tabs extends React.Component<P.TabsProps> {}
