+++ types/muicss/lib/react/textarea.d.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import * as P from '../../react';
+
+export default class TextArea extends React.Component<P.TextareaProps> {}
